src/backend/requirements.txt
Django>=1.8,<2.0
gunicorn==19.6.0
psycopg2
djangorestframework
markdown
djangorestframework-jwt